Refactor: 크루 시작시간, 종료시간 자료형 변경 (Integer -> Double)

+++ src/main/java/com/server/wupitch/club/Club.java
@@ -69,9 +69,9 @@ public class Club extends BaseTimeEntity {
 
     private Boolean sunday;
 
-    private Integer startTime;
+    private Double startTime;
 
-    private Integer endTime;
+    private Double endTime;
 
     private Long memberCount;
